// app/model/serverinfoSummary-model.ts
export class ServerinfoSummary {
        node: string;
        public_ip: string;
        private_ip: string;
        fqdn: string;
        domain_name: string;
         operating_system: string;
        data_center: string;
        provider: string;
        project: string;
        role: string;
    }